requirements.txt
asgiref==3.6.0
infobip-api-python-sdk
boto3==1.28.27
botocore==1.31.27
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-database-url==0.4.2
Django==4.2
django-cors-headers==3.14.0
django-ses==3.5.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
environ==1.0
filelock==3.12.2
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
sympy==1.12
typing_extensions==4.7.1
urllib3==1.26.16
whitenoise==6.5.0
